+++ src/System/Database.php
@@ -19,7 +19,14 @@ class Database {
 
     use SingletonTrait;
 
-    private $files, $event, $io, $config;
+    /** @var Files $files */
+    private $files;
+    /** @var Event $event */
+    private $event;
+    /** @var IOInterface $io */
+    private $io;
+    /** @var Data $config */
+    private $config;
 
 
     /**
@@ -30,7 +37,7 @@ class Database {
     public function __construct(Event $event)
     {
 
-        $this->files = new Files();
+        $this->files = new Files( $event );
         $this->event = $event;
         $this->io    = $event->getIO();
 
@@ -52,6 +59,7 @@ class Database {
 
         foreach ( [
                       'global',
+                      'remote',
                       'local'
                   ] as $config ) {
 
@@ -65,71 +73,250 @@ class Database {
         $this->config = new Data( $data );
     }
 
+    public static function handle(Event $event)
+    {
+        /** @var Database $database */
+        $database = Database::getInstance( $event );
+        $args     = $event->getArguments();
+
+        if ( !count( $args ) ) {
+
+            $database->io->writeError( $database->getComposerDatabaseDescription() );
+
+            return;
+        }
+
+        $action = $args[0];
+
+        if ( $action != 'import' && $action != 'export' ) {
+
+            $database->io->writeError( "  Wrong action call\n" . "  action can be 'import' or 'export' only." . $database->getComposerDatabaseDescription() );
+
+            return;
+        }
+
+        if ( $action == 'import' && count( $args ) < 2 ) {
+
+            $database->io->writeError( "  Missing path argument\n" . $database->getComposerDatabaseDescription() );
+
+            return;
+        }
+
+        $archive_info = $args[1];
+
+        try {
+
+            $database->prepare( $action, $archive_info );
+
+        } catch ( \Exception $e ) {
+
+            $database->io->write( "  ERROR : " . $e->getMessage() );
+        }
+    }
+
+    /**
+     * @param $action
+     * @param $archive_info
+     */
+    public function prepare($action, $archive_info)
+    {
+
+        if ( !method_exists( $this, $action ) ) {
+
+            throw new \InvalidArgumentException( $action . " function does not exist !" . $this->getComposerDatabaseDescription() );
+        }
+
+        $db_infos = [];
+        // Common configuration
+        $db_infos['user']     = $this->config->get( 'database.user' );
+        $db_infos['password'] = $this->config->get( 'database.password' );
+        $db_infos['name']     = $this->config->get( 'database.name' );
+        $db_infos['host']     = "localhost";
+
+        $this->io->write( '  Starting ' . $action . '...' );
+
+
+        return $this->$action( $archive_info, $db_infos );
+    }
+
+
+    /**
+     * Process Withdrawing from remote configuration to local
+     *
+     * @param $remote_cfg
+     */
+    public function withdraw($remote_cfg)
+    {
+        $tmp_dir = 'var';
+
+        $replacements = isset( $remote_cfg['database']['replace'] ) ? $remote_cfg['database']['replace'] : false;
+
+        // Creating local informations
+        $local_cfg = $this->config->get('database');
+
+        if( !isset($local_cfg['host']) )
+            $local_cfg['host'] = "localhost";
+
+        $exported_db = $this->export( $tmp_dir, $remote_cfg['database'], $replacements );
+
+        $this->import( $exported_db, $local_cfg );
+
+        unlink( $exported_db );
+    }
+
+
+    /**
+     * Process deployment of local configuration to given remote configuration
+     *
+     * @param $remote_cfg
+     */
+    public function deploy($remote_cfg)
+    {
+        $tmp_dir = 'var';
+
+        if ( isset( $remote_cfg['database']['replace'] ) )
+        {
+            // deployment is the reverse replacement array
+            $replacements = [];
+            foreach ( $remote_cfg['database']['replace'] as $replacement )
+            {
+                foreach ( $replacement as $new => $old )
+                {
+                    $replacements[] = [$old => $new];
+                }
+            }
+        }
+        else
+        {
+            $replacements = false;
+        }
+
+        // Creating local informations
+        $local_cfg = $this->config->get('database');
+
+        if( !isset($local_cfg['host']) )
+            $local_cfg['host'] = "localhost";
+
+        $exported_db = $this->export( $tmp_dir, $local_cfg, $replacements );
+
+        $this->import( $exported_db, $remote_cfg['database'] );
+
+        unlink( $exported_db );
+    }
+
 
     /**
      * Import database
      */
-    public static function import(Event $event)
+    public function import($archive_path, $db_infos)
     {
+        if ( !is_array( $db_infos ) || !isset( $db_infos['user'], $db_infos['password'], $db_infos['name'], $db_infos['host'] ) )
+        {
+            throw new \InvalidArgumentException( "Database informations are not complete." );
+        }
 
-        $app_path = getcwd() . DIRECTORY_SEPARATOR . "app";
-        $database = Database::getInstance( $event );
-        $args     = $event->getArguments();
+        if ( file_exists( $archive_path ) )
+        {
+            $confirm = $this->io->askConfirmation( "  File name : " . $archive_path . "\n  Confirm import ? [y,n] ", false );
 
-        if ( count( $args ) ) {
-            $filename = $app_path . '/backup/' . $args[0];
+            if ( $confirm )
+            {
+                $file_info = pathinfo( $archive_path );
+
+                $command = "";
+
+                // extracting file content
+                if ( $file_info['extension'] == 'gz' )
+                {
+                    $command .= "zcat " . $archive_path . " | ";
+                }
+
+                $command .= "mysql -u " . $db_infos['user'] . " -h " . $db_infos['host'] . " --password='" . $db_infos['password'] . "' " . $db_infos['name'];
+
+                passthru( $command );
+
+                $this->io->write( "\n  Import complete." );
+            }
         }
         else {
-            $filename = $app_path . '/resources/db.sql';
-        }
 
-        if ( file_exists( $filename ) ) {
-
-            $database->io->write( '  Importing database...' );
-
-            if ( count( $args ) == 3 ) {
-
-                file_put_contents( $filename . '.tmp', str_replace( $args[1], $args[2], file_get_contents( $filename ) ) );
-                $filename = $filename . '.tmp';
-            }
-
-            passthru( "mysql -u " . $database->config->get( 'database.user' ) . " -p" . $database->config->get( 'database.password' ) . " " . $database->config->get( 'database.name' ) . " < " . $filename . "  2>&1 | grep -v \"Warning: Using a password\"" );
-            $database->io->write( '  Import complete' );
-
-            if ( count( $args ) == 3 ) {
-                unlink( $filename );
-            }
-        }
-        else {
-
-            $database->io->write( '  ' . $filename . ' does not exists' );
+            throw new \InvalidArgumentException( $archive_path . ' does not exists.' );
         }
     }
 
 
     /**
      * Export database
+     *
+     * @param $export_dir string path to exported database
+     * @param $db_infos   array database infos
+     * @param bool $replacements
+     * @return string filename
      */
-    public static function export(Event $event)
+    public function export($export_dir, $db_infos, $replacements = false)
     {
-
-        $database = Database::getInstance( $event );
-        $backup_path = getcwd() . DIRECTORY_SEPARATOR . "app/backup";
-        $args = $event->getArguments();
+        if ( !$export_dir )
+        {
+            $backup_path = getcwd() . DIRECTORY_SEPARATOR . "app/resources/db";
+        }
+        else
+        {
+            $backup_path = getcwd() . DIRECTORY_SEPARATOR . $export_dir;
+        }
 
         if ( !is_dir( $backup_path ) ) {
             mkdir( $backup_path );
         }
 
-        $filename = $backup_path . '/' . date( 'Ymd' ) . '.sql';
+        $filename = $backup_path . '/' . date( 'Ymd' ) . '.sql.gz';
 
-        $database->io->write( '  Exporting database...' );
-        passthru( "mysqldump -u " . $database->config->get( 'database.user' ) . " -p" . $database->config->get( 'database.password' ) . " " . $database->config->get( 'database.name' ) . " > " . $filename );
+        $this->io->write( '  Exporting database...' );
+        $dump        = "mysqldump -u " . $db_infos['user'] . " --password='" . $db_infos['password'] . "' -h " . $db_infos['host'] . " " . $db_infos['name'];
+        $compression = "gzip > " . $filename;
+        $replace     = ( $replacements ? ' | ' . $this->getReplace( $replacements ) : '' );
 
-        if ( count( $args ) == 2 ) {
-            file_put_contents( $filename, str_replace( $args[0], $args[1], file_get_contents( $filename ) ) );
+        $command = $dump . $replace . ' | ' . $compression;
+
+        passthru( $command );
+
+        $this->io->write( "  Exporting successful !\n" . "  Path to export : " . $filename );
+
+        return $filename;
+    }
+
+
+    /**
+     * Replace database strings to new strings
+     *
+     * @param $replacements array
+     * @return string thh command
+     */
+    private function getReplace($replacements)
+    {
+        $command = '';
+
+        if ( isset( $replacements ) )
+        {
+            $command = "sed '";
+            foreach ( $replacements as $replacement )
+            {
+                foreach ( $replacement as $old => $new )
+                {
+                    $command .= "s," . $old . "," . $new . ",g; ";
+                }
+            }
+            $command .= "' ";
         }
 
-        $database->io->write( '  Exporting complete' );
+        return $command;
+    }
+
+
+    /**
+     * @return string
+     */
+    public function getComposerDatabaseDescription()
+    {
+        return "\n  ----------------- \n" . "  COMPOSER DATABASE \n" . "  ----------------- \n" . "  composer database [action] [path]\n" . "           [action] : \n" . "             - import\n" . "             - export\n" . "           [path] : relative path to directory or database\n";
     }
 }